# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "OAuth 2 based token issue and validation server, with built in management UI",
        "title": "Authserver API",
        "termsOfService": "https://github.com/danesparza/authserver",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/danesparza/authserver"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/danesparza/authserver/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/",
    "paths": {
        "/oauth/authorize": {
            "get": {
                "security": [
                    {
                        "OAuth2Application": []
                    }
                ],
                "description": "gets the scope information for the userID passed in the url",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "gets the scope information",
                "operationId": "scopes-for-user-id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.AuthResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.AuthResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string"
                }
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://localhost:3001/oauth/token/client",
            "scopes": {
                "sys_admin": " Grants read and write access to administrative information",
                "sys_delegate": " Grants write access for a specific resource"
            }
        }
    }
}